+++ castle/cms/interfaces/__init__.py
@@ -16,6 +16,7 @@ from .controlpanel import IAPISettings
 from .controlpanel import IArchivalSettings
 from .controlpanel import IBusinessData
 from .controlpanel import ICastleSettings
+from .controlpanel import IContentSettings
 from .controlpanel import ICrawlerConfiguration
 from .controlpanel import ISecuritySchema
 from .controlpanel import ISiteConfiguration