+++ castle/cms/pwexpiry/password_history_validator.py
@@ -1,10 +1,9 @@
 from AccessControl import AuthEncoding
-from castle.cms.pwexpiry.interfaces import ICustomPasswordValidator
+from castle.cms.interfaces.passwordvalidation import ICustomPasswordValidator
 from plone import api
 from zope.interface import implementer
 
 from .config import _
-from .interfaces import ICollectivePWExpiryLayer
 
 
 @implementer(ICustomPasswordValidator)
@@ -21,12 +20,6 @@ class PasswordHistoryValidator(object):
         Password validation method
         """
 
-        # Don't check the password if castle.cms.pwexpiry hasn't
-        # been installed yet.
-        if not ICollectivePWExpiryLayer \
-                .providedBy(self.context.REQUEST):
-            return None
-
         # Permit empty passwords here to allow registration links.
         # Plone will force the user to set one.
         if password is None:
@@ -35,12 +28,6 @@ class PasswordHistoryValidator(object):
         if api.user.is_anonymous():
             # No check for registrations.
             return None
-
-        if len(password) < 8:
-            return _(
-                u'Minimum 8 characters',
-                default=u'Minimum 8 characters'
-            )
 
         pwexpiry_enabled = api.portal.get_registry_record(
             'plone.pwexpiry_enabled'